# Optimize onehot2int for performance