## Summary
This PR partially addresses #862 

> 🎯 **Purpose**: Optimize onehot2int for performance.
>

**[ ✏️ Write your summary here. ]**
While profiling another function I saw that the list creation was taking a significant time in this function. By using numpy method we can speed this function significantly.

For memory I used the memory-profiler library. The code I used for benchmarking is copied below. In addition I sorted the imports in the modified files.

**Code Setup**
```python
N = 10_000_000
K = 20
onehot_matrix = np.zeros((N, K), dtype=int)

# Fill some random 1s in the onehot_matrix
for _ in range(3):
    index = np.random.randint(K, size=N)
    onehot_matrix[np.arange(N), index] = 1
```

**Current version**
```python
%%timeit
%memit onehot2int(onehot_matrix)
# peak memory: 3037.66 MiB, increment: 1140.43 MiB
# peak memory: 3038.42 MiB, increment: 1138.13 MiB
# peak memory: 3041.00 MiB, increment: 1140.24 MiB
# peak memory: 3036.71 MiB, increment: 1135.98 MiB
# peak memory: 3038.43 MiB, increment: 1137.52 MiB
# peak memory: 3037.65 MiB, increment: 1137.47 MiB
# peak memory: 3040.58 MiB, increment: 1140.40 MiB
# peak memory: 3038.36 MiB, increment: 1139.16 MiB
# 16.3 s ± 109 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


**This PR**
```python
%%timeit
%memit onehot2int(onehot_matrix)
# peak memory: 2828.12 MiB, increment: 952.94 MiB
# peak memory: 2821.07 MiB, increment: 961.35 MiB
# peak memory: 2817.46 MiB, increment: 956.20 MiB
# peak memory: 2827.25 MiB, increment: 965.52 MiB
# peak memory: 2819.76 MiB, increment: 957.07 MiB
# peak memory: 2825.89 MiB, increment: 963.88 MiB
# peak memory: 2821.27 MiB, increment: 960.11 MiB
# peak memory: 2820.75 MiB, increment: 958.90 MiB
# 5.94 s ± 49.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


## Testing

> 🔍 Testing Done: Existing test suite and I also verified that the outputs were the same after refactoring.


## References


## Reviewer Notes
> 💡 Include any specific points for the reviewer to consider during their review.